+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Add WebConfigMQTT class and replace MQTTClient with it.
+* Fix getMQTTClient client pointer retrun and enable loop method only  for ESP8266 target
 
 v2.0.1 (2023-05-08)
 ------